refactored Makefile to account for new compilation scenario of compiling ftcontainers only when invoking tests

+++ Makefile
@@ -1,41 +1,20 @@
-STATIC_LIB = libftcontainers.a
-
 CC = c++ -Wall -Wextra -Werror -std=c++98
 
 BINARIES_DIR = bin
 HEADERS_DIR = include
 TEMPLATES_DIR = src
-OBJS_DIR = objs
 TESTS_DIR = tests
 
-HEADERS = ft_map.hpp ft_set.hpp ft_stack.hpp ft_vector.hpp
-TEMPLATES = ft_map.tpp ft_set.tpp ft_stack.tpp ft_vector.tpp
+HEADERS = $(addprefix $(HEADERS_DIR)/,	ft_vector.hpp)
+TEMPLATES = $(addprefix $(TEMPLATES_DIR)/,	ft_vector.tpp)
 
-OBJECTS = $(addprefix $(OBJS_DIR)/,$(HEADERS:.hpp=.o))
+TEST_FILES =	$(addprefix $(TESTS_DIR)/,	main.cpp test_emptiness.cpp test_typedefs.cpp)
 
-TEST_FILES =	$(addprefix $(TESTS_DIR)/,	main.cpp \
-											test_typedefs.cpp \
-											test_vector_creation.cpp)
+all: tests
 
-all: $(STATIC_LIB)
-
-$(STATIC_LIB): $(OBJS_DIR) $(OBJECTS)
-	ar rcs $(STATIC_LIB) $(OBJECTS)
-
-$(OBJS_DIR)/%.o: $(HEADERS_DIR)/%.hpp
-	$(CC) $(CFLAGS) -c -I $(TEMPLATES_DIR) $< -o $@
-
-$(OBJS_DIR):
-	if [ ! -d "$(OBJS_DIR)" ]; then mkdir $(OBJS_DIR); fi
-
-run: $(STATIC_LIB)
-	if [ ! -d "$(BINARIES_DIR)" ]; then mkdir $(BINARIES_DIR); fi
-	$(CC) $(CFLAGS) -I $(HEADERS_DIR) -I $(TEMPLATES_DIR) main.cpp -o $(BINARIES_DIR)/main.out
-	./$(BINARIES_DIR)/main.out
-
-tests: $(STATIC_LIB)
-	$(CC) $(CFLAGS) -I $(HEADERS_DIR) -I $(TESTS_DIR)/include -I $(TEMPLATES_DIR) $(TEST_FILES) -o $(BINARIES_DIR)/tests.out
-	./$(BINARIES_DIR)/tests.out
+tests: $(HEADERS) $(TEMPLATES) $(TEST_FILES)
+	$(CC) $(CFLAGS) -I $(HEADERS_DIR) -I $(TESTS_DIR)/include -I $(TEMPLATES_DIR) $(TEST_FILES) -o tests.out
+	./tests.out
 
 valgrind: all
 	valgrind --tool=memcheck \
@@ -49,14 +28,13 @@ valgrind: all
 		./$(BINARIES_DIR)/main.out
 
 clean:
-	rm -f $(OBJECTS)
-	rmdir --ignore-fail-on-non-empty $(OBJS_DIR)
+# 	rm -f $(OBJECTS)
+	if [ -d "$(OBJS_DIR)" ]; then rmdir --ignore-fail-on-non-empty $(OBJS_DIR); fi
 
 fclean: clean
 	rm -f $(BINARIES_DIR)/*
-	rmdir --ignore-fail-on-non-empty $(BINARIES_DIR)
-	rm -f $(STATIC_LIB)
+	if [ -d "$(BINARIES_DIR)" ]; then rmdir --ignore-fail-on-non-empty $(BINARIES_DIR); fi
 
 re: fclean all
 
-.PHONY: all clean fclean re run tests valgrind
+.PHONY: all clean fclean re tests valgrind
